requirements.txt
kafka-python
requests
pandas
python-dotenv
snowflake-connector-python
